+++ src/Recyclarr.Cli/Pipelines/IPipelineMetadata.cs
@@ -1,4 +1,5 @@
 using Recyclarr.Sync;
+using Recyclarr.TrashGuide;
 
 namespace Recyclarr.Cli.Pipelines;
 
@@ -6,4 +7,5 @@ internal interface IPipelineMetadata
 {
     static abstract PipelineType PipelineType { get; }
     static abstract IReadOnlyList<PipelineType> Dependencies { get; }
+    static abstract SupportedServices? ServiceAffinity { get; }
 }