+++ src/Recyclarr.Cli/Pipelines/MediaManagement/MediaManagementPipelineContext.cs
@@ -1,5 +1,6 @@
 using Recyclarr.Servarr.MediaManagement;
 using Recyclarr.Sync;
+using Recyclarr.TrashGuide;
 
 namespace Recyclarr.Cli.Pipelines.MediaManagement;
 
@@ -7,6 +8,7 @@ internal class MediaManagementPipelineContext : PipelineContext, IPipelineMetada
 {
     public static PipelineType PipelineType => PipelineType.MediaManagement;
     public static IReadOnlyList<PipelineType> Dependencies => [];
+    public static SupportedServices? ServiceAffinity => null;
 
     public override string PipelineDescription => "Media Management";
     public override bool ShouldSkip => !Plan.MediaManagementAvailable;