+++ src/Recyclarr.Cli/Pipelines/QualitySize/QualitySizePipelineContext.cs
@@ -1,6 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
 using Recyclarr.Servarr.QualitySize;
 using Recyclarr.Sync;
+using Recyclarr.TrashGuide;
 using Recyclarr.TrashGuide.QualitySize;
 
 namespace Recyclarr.Cli.Pipelines.QualitySize;
@@ -14,6 +15,7 @@ internal class QualitySizePipelineContext : PipelineContext, IPipelineMetadata
 {
     public static PipelineType PipelineType => PipelineType.QualitySize;
     public static IReadOnlyList<PipelineType> Dependencies => [];
+    public static SupportedServices? ServiceAffinity => null;
 
     public override string PipelineDescription => "Quality Definition";
     public override bool ShouldSkip => !Plan.QualitySizesAvailable;