+++ src/Recyclarr.Core/Notifications/NotificationService.cs
@@ -221,8 +221,7 @@ internal sealed class NotificationService : INotificationService, IDisposable
             PipelineType.CustomFormat => "Custom Formats Synced",
             PipelineType.QualityProfile => "Quality Profiles Synced",
             PipelineType.QualitySize => "Quality Sizes Synced",
-            PipelineType.SonarrMediaNaming => "Media Naming Synced",
-            PipelineType.RadarrMediaNaming => "Media Naming Synced",
+            PipelineType.MediaNaming => "Media Naming Synced",
             _ => "Items Synced",
         };
     }